+++ man-roxygen/params-icd-year.R
@@ -1,4 +1,5 @@
 #' @param year Numeric scalar. Calendar or fiscal year to reference. Default
-#'   is the most current year available per source. For ICD-9, the latest year
-#'   is 2015; ICD-10 source are updated annually.  Calendar year for WHO and CDC
-#'   mortality. Fiscal year for CMS.
+#'   is the most current year available per source. For ICD-9, CMS data run
+#'   through fiscal year 2015 and CDC extracts through 2012; ICD-10 sources are
+#'   updated annually. Calendar year for WHO and CDC mortality. Fiscal year for
+#'   CMS.